move fee event recording to tx level
